requirements.txt
# Software Defect Prediction Pipeline - Dependencies
# Python 3.8+ required

# Core numerical computing
numpy>=1.21
pandas>=1.3

# Machine Learning
scikit-learn>=1.0
xgboost>=1.5
torch>=2.0

# Class imbalance handling
# Custom SMOTE implementation included (src/preprocessing.py) for maximum compatibility
# Uncomment below if using Python <= 3.12 with sklearn <= 1.5
# imbalanced-learn>=0.9

# Explainability
lime>=0.2

# Visualization
matplotlib>=3.5
seaborn>=0.11

# Statistical testing
scipy>=1.7

# ARFF file parsing
liac-arff>=2.5

# Progress display
tqdm>=4.60

